consistent lookup: handle insert ignore
Fixes #4974

Insert ignore wasn't working correctly because the verify was using
the normal transaction instead of the 'pre' transaction where the
rows were actually created.

This implementation changes the consistent lookup's Verify function
to also use the 'pre' transaction always.

There is another code path that gets used if the vindex is unowned.
That will also cause a 'pre' transaction to be created. Without this
change, the transaction would have been a normal one. This mildly
affects the commit order, but there should be no material difference.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>