vcopier: use ticker instead of timer
The vcopier was using a timer instead of a ticker in its wait loop
which caused it to wait forever even if the replication had caught up.

The change includes a new test that tests the full copy cycle and
ensures that multiple iterations succeed.

A similar incorrect usage was found in logutil. That's also been
fixed.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>