Allow start local cluster using a provided tablet host name
This helps fix a name resolution problem when initializing the tablet. If flag "tablet_hostname" is not provided it will be derived from the OS' hostname. The way it is implemented there is not guarantee that ```net.LookupHost(os.HostName())``` will return a valid IP for certain local env dns configs, i.e, in Mac OS X, hostname can be configured using the system's resolver which is not used by ```net.LookupHost```

Reference: init_tablet.go:151

Signed-off-by: Karel Alfonso Sague <kalfonso@squareup.com>