vreplication: bug: make row streaming more robust
During production testing, we found that there are possibilities
that a table lock never gets released when row streaming is requested.

We most likely suspect that a grpc send hangs, which keeps the table
locked, when eventually causes replication to stall and subsequent
row streaming requests to hang.

This change refactors the code to make sure that we unlock a table
as soon as streaming has started. The previous code unlocked the table
only after sending the field info.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>