support more expression types in vtexplain
Rework the vtexplain simulated mysql to handle more expression types
when calculating fake column results based on inferences off the select
expression.

Also change the engine so that it logs an error instead of failing
outright, since the column values aren't always needed for it to
effectively simulate a query.

Signed-off-by: Michael Demmer <mdemmer@slack-corp.com>