vtbackup: Make initial backup idempotent again.
The check for a complete backup already existing was supposed to make
the initial backup idempotent, such that running it after an initial
backup already exists will simply do nothing and return success.

However, that property was broken when I added the check for serving
tablets above the check for a complete backup. I should have kept the
check for a complete backup first, because if there's already a complete
backup and we're doing nothing, then it's fine that some tablets may
already be serving.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>