Handle case where mysqld replies to Initial Handshake Packet with an ERR packet.
From mysql docs:

```
It starts with the client connect()ing to the server which may send a
ERR packet and finish the handshake or send a Initial Handshake
Packet...
```

`sql_state` fields are skipped because they are not set according to docs.

```
Note
In case the server sent a ERR packet as first packet it will happen
before the client and server negotiated any capabilities. Therefore
the ERR packet will not contain the SQL-state.
```

Signed-off-by: Adam Saponara <as@php.net>