replication_reporter: Don't try to reparent to yourself.
We've seen it happen that a master tablet restarts and becomes a
replica. If the shard record still says we are master, we might end up
trying to reparent to ourselves.

I don't know yet how the tablet is getting forced to replica type, but
in any case we should enforce the invariant that we don't try to
reparent to ourselves.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>