vtbackup: Retry fetching master position.
We've seen that this could get stuck if vtbackup happens to read the
master tablet record while the tablet is unreachable, and then the
tablet comes back up with a different address. This adds a retry that
re-reads the shard and tablet records to pick up any changes in the
master alias or tablet address.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>