xtrabackup: Add a timeout on closing backup files.
We've seen backup attempts that apparently stalled while waiting for
Close() on the file returned by AddFile() to return. We've only seen
this on xtrabackup backups, likely because we perform a small number of
long-running file uploads, instead of uploading each file individually.

This adds a timeout to the Close() step. If it times out, the backup
will be aborted and will need to be retried from scratch. However,
that's better than getting stuck forever.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>