topo: modify CreateShard IsMasterServing rule
CreateShard starts off all unsharded shards as serving. It
breaks if you create shards -80, 80-, and then 0
(with the intention to reshard to 0).

It was previously allowed because we allowed the old style of custom
sharding where you could create shards 0, 1, 2, etc.
But we don't support that any more.

The new rule sets master as serving only if the keyranges don't
overlap. This is applied for unsharded keyspaces also.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>