engine: simplify bind var initialization
Bind var initialization was duplicated in different parts of
the primitive. It's now unified and done after all the vindexes
are processed.

Also, the primary vindex code was flattening the vindex values
because it assumed only one column. The code has been rewritten
to not assume this.

This prepares us for the next step, which is to implement MapNew
that will accept multi-column vindex values.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>