Don't recheck mysqld port if it was provided explicitly.
At startup, we already avoid asking mysqld for its port if we were given
the port in DB configs. However, we would forget this port when we later
trigger a recheck of the MySQL port.

It's important to trust the port that was given to us because it might
be different from the port that mysqld is actually serving on, such as
in the case that the host we were given is actually a proxy.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>