vrepl: handle "other" events
While working on filePos flavor, I encountered this issue where
an "other" statement will cause the next GTID to not be immediately
sent. This can cause delays if the target waits for that event.

This is pretty rare for GTID mode. But will likely be more pronounced
for filePos. So, I'm proactively making this fix to make sure we
don't delay sending of GTIDs, even in the case of non-relevant events.

Because of this change, this also means that we don't have to generate
pseudo-gtids. Generating GTIDs outside of transactions, coupled with
OTHER event will make the right thing happen.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>