vreplication: standardize on stop pos meaning
The vplayer currently uses ambiguous rules about how it handles the case
where a stop position was exceeded. As part of this change, we'll
standardize on: A stop position is considered to be successfully reached
if the new position is greater than or equal to the specified position.
The main motivation for this change is that the possibility of
position mismatch is higher in the case of file:pos tracking. We're
likely to hit many false positives if we're too strict.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>