Allow to demote master vttablet to SPARE instead of REPLICA
In our setups we use SPARE tablets to take over MASTER role during reparents or failures. Currently shard sync and external reparent demote MASTER to REPLICA. This change allows to configure the tablet type of a demoted master. SPARE and REPLICA are the only ones allowed. REPLICA is the default one if flag is not set. This can also be achieved with a boolean flag if only SPARE and REPLICA make sense as the only tablet types that can be promoted to master

Signed-off-by: Karel Alfonso Sague <kalfonso@squareup.com>