vrepl: improve row and result streamers
The new implementation starts a transaction after locking the
tables instead of issuing the read query. This guarantees that
we'll release the table locks immediately. Otherwise, there's
no guarantee about when the first results will be returned.

Also, the logic to do the snapshotting has been moved into
a reusable connection wrapper, which is a more elegant approach.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>