Fix vreplication panic
Fixes #5752

DBClient was setting the connection to nil after close, which was
causing vreplication to panic because it defers a Rollback on all
connections. The safer thing is to close the connection, but not
set it to nil.

There's no test for DBClient, but I've manually verified that this
won't panic any more.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>