Remove ParenExp from code base
Instead of keeping this information in the AST, we put in parens
in the string representation of queries when needed when printing.

This includes adding code that takes precendence into consideration - something
that already we already need to do when parsing. Unfortunately I haven't found
a good way to use the same source information for both sides of the
precedence puzzle.

Signed-off-by: Andres Taylor <andres@planetscale.com>