vreplication: fix bug on stop position
Fixes #5800

The function hasAnother commit is supposed to separate out events
that have to be applied in a transaction from events that are
applied as autocommits. But it was not checking for the OTHER
and JOURNAL events. They should not be batched with regular
transactions.

This caused a bug where a regular transaction occured followed
by an OTHER event. The apply of this event happened assuming
autocommit behavior, but it actually got added to the previous
uncommitted transaction.

If the stop position is reached at this point, vplayer exits
because it thinks the autocommit happened, and the entire
transaction gets rolled back, and the stop position ends
up not being actually reached.

The copier which expects this behavior will then start applying
the next set of rows, but they are not consistent with the current
stop position. This will cause the target to go out of sync.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>