vstreamer: bug: use WithCredentials
VStreamer was not prefacing mysql.Connect calls with WithCredentials.
I've now added the calls.

There are no new tests for this because this is not a sustainable
coding pattern. I'll create a separate issue to refactor the code
such that one should not have to remember to call WithCredentials
before every call to mysql.Connect.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>