vrepl: vplayer must rollback on exit
There were code paths where vplayer had an open transaction on
its dbclient connection, which got accidentally continued by
the copier. This caused production issues with some users.

The specific use case that was observed is:
* vreplication performs catchup.
* While it's in the middle of an apply, the context is canceled.
* The copier has no work.
* vreplication is asked to resume from the last saved position.
* It replays the same statements already in the transaction.
* This leads to dup key errors.

This fix conservatively rolls back dbclient before exiting.
The dbclient itself skips rollback if it's not in a transaction.
This change leads to relieable rollbacks where needed. At the
same time, there are no spurious rollbacks if we didn't start
a transaction.

As safety, I've added an extra rollback in the vreplicator loop
just in case the underlying functions accidentally leave an
incomplete transaction open.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>