Instead of failing when trying to skip updates, fallback on slower mode
When updating consistent lookup vindexes, we see if anything really did change,
before updating the vindex. If nothing changes, we can skip updating the vindex.

Sometimes, we encounter errors when checking if we can shortcut the updates.
Instead of passing on there errors, we'll now fall back on the slower mode and execute the update.

Signed-off-by: Andres Taylor <andres@planetscale.com>