vstreamer: send immediate GTID on "current"
If the current position is requested as start for vstream, it will
be good to send the actual starting GTID as a separate event before
streaming the rest.

This is particularly useful for messager because it needs to know
its current position from the beginning in order to determine which
events to skip w.r.t to poller position.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>