messager: removed waitgroup on vstream
Adding a waitgroup for vstream caused all kinds of deadlocks
because vstreamer calls se.Reload, which calls back into engine
on schemaChanged.

stopVStream is already safe: it will eventually cancel the VStreamer.
Any call back after stop will be a no-op. So, there's not much
value in waiting for that goroutine to return before declaring
Close as done.

So, not waiting on vstream to exit before closing makes things
more stable.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>