Refactoring of Myprintf
Myprintf uses from AST now include the parent struct.

In order to get better readability and guarantees that we don't introduce regressions in the future,
this refactoring changes Myprintf uses from the AST to always provide the parent structure. Using
this information, Myprintf can figure out when parens are needed.

Signed-off-by: Harshit Gangal <harshit@planetscale.com>
Signed-off-by: Andres Taylor <andres@planetscale.com>