vtgate send primitive: implement StreamExecute
In #5982 queries after `use keyspace[shard]` were changed to use a new "Send" engine primitive instead of v2 routing, which caused OLAP to stop working for queries targeted in that way. This PR fixes those OLAP queries.

This PR does not implement GetFields, so prepared statements still aren't going to work with this kind of targeting.

As a bonus this also fixes an issue where trying to start/rollback/commit a transaction in OLAP mode caused a panic due to plan.Instructions being nil in that case.

Signed-off-by: David Weitzman <dweitzman@pinterest.com>