vttablet: fix panic in NewDBConnNoPool
Fixes #6084.

The last exporter change introduced a regression due to NewDBConnNoPool
creating a tabletenv.Stats with a named exporter, which conflicts with
the variables created by the unnamed exporter. This change reuses
the stats created for the tabletserver instead.

I've tested that the old code panics, and that it doesn't panic after
the fix. But I removed the test because it can create other problems
if someone created unnamed stats elsewhere.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>