Remove synchronizing VTGateConnection on itself
We're doing some feature work on a fork of Vitess where this
synchronized instance lock is creating extra pausing in our client
application. The only state the connection maintains is the RpcClient,
which GRPC should not have a need to be synchronized.

We've removed this blocks in our fork and run some overall query load
testing (nothing explicitly targeting this code block, just throwing a
bunch of transaction reads and inserts through the client). We didn't
encounter any noticeable issues with the change.

Reading the code, the only thing that seems stateful to me is the
VtSession maybe racing between the check to allow a call and then
setting the `lastCall` after constructing the new `SQLFuture`. However,
I don't think an instance lock on the `VTGateConnection` would prevent
anyone else from racing against the `VTSession` instance? I do not have
complete understanding, this is just my surface level observation.

Signed-off-by: Rich Marscher <rmarscher@hubspot.com>