vtctld: make cell optional
This change makes the cell flag optional for vtctld. This removes
the catch-22 that you have to create a cell before launching vtctld.
But then, how do you create a cell without a vtctld?

This required a one-time use of vtctl tool to create this initial
cell.

If you don't specify a cell, vtctld will choose one at random.
I believe the main use case for this is for identifying which shards
are serving vs. non-serving. This info is now available in global
topo itself as the "IsMasterServing" flag. But, we'll need to
go into the UI to make this change, which is non-trivial right now.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>