go/mysql: Fix MariadbGTIDSet multi-domain support.
This fixes two bugs:

1. AddGTID was mutating the original set. GTIDSet implementations are
supposed to be immutable.

2. AddGTID and parseMariadbGTIDSet did not enforce any order among
domains, yet Equal and other functions expected order to be enforced.

This switches from a list to a map to better represent the fact that the
order of domains doesn't matter, which also simplifies some code.
We always enforce order when serializing back to a string.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>