Make it easier to test multiple endpoints with aya
This cleans up the interface of are you alive to pass it a config file
with endpoint configuration rather than command line options that make
are you alive construct the connection strings in different ways.

This makes it easier to test a larger number of endpoints, so that, for
example, you could easily configure it to write to one endpoint while
reading from 11 others, at any rate you want.

Also removes environment from the labels since that could be added later
by the deployment, and makes the rate configuration a true rate instead
of a delay.

Signed-off-by: Shaun Verch <sverch@planetscale.com>