Simplify vtgate executor
We introduced an abstraction layer to be able to have the old and new executor
side by side in the same code base. We never really used this capability and
instead have been moving logic steadly from one to the other, and then removing
the legacy code.

Given this situation, the extra abstraction was just awkward to explain, and not at all helpful.

This simplifies the code by removing unecessary abstractions and tests.

Signed-off-by: Andres Taylor <andres@planetscale.com>