tm revamp: deprecate demoteMasterType
No one knows this exists, and no one uses it. The intuitive tablet
type is the one from init_tablet_type, which everybody already sets
while starting a vttablet.

Because this is the only use case in real life, this change also
requires init_tablet_type to be always specified.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>