vttablet: fix deadlock in schema engine
The new historian reuses the pool from schema engine. But that pool
has only one connection. I found a deadlock where se calls the
historian with that conn open, and the historian in turn requests
a connection for its own reload. This causes a deadlock. This is
fixed by increasing pool size to 2.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>