[PATCH] stream_query_plan_cache.patch
adds a query plan cache to OLAP mode for parity with OLTP mode.

We wrote this while trying to improve perf. It seemed to help a
bit. Haven't perf tested this in a while.

Signed-off-by: Adam Saponara <as@php.net>

Signed-off-by: mstarr <mstarr@etsy.com>
Signed-off-by: Adam Saponara <as@php.net>

Add feature parity with query plans; add query_engine tests

Signed-off-by: Samantha Drago-Kramer <skramer@etsy.com>

Use a single cache for stream and non-stream query plans

Signed-off-by: Samantha Drago-Kramer <skramer@etsy.com>