Add `-rename_tables` to DropSources
This addition allows the optional bool -rename_tables for DropSources.
This can be useful when dropping sources for larger tables. Instead of a
`drop table`, a `rename table` is run. This will be faster in most cases
but will require the user to clean up the tables at a later date.

Signed-off-by: Tom Krouper <tomkrouper@github.com>