mysql/client: Initial handshake should not validate auth plugin
During the initial handshake, we are checking if the default auth plugin
of the server is supported by the client. In cases where the server
supports multiple auth plugins, this check is causing the handshake to
fail without giving the server a chance to send an Authentication Method
Switch Request.

This check happens anyway when handling the auth method switch request
and an error is thrown when server requests an unsupported auth method.

Signed-off-by: Vamsi Atluri <vamc19@gmail.com>