tx limiter: fix quota leak if Begin() errors or if the tx killer runs
The transaction limiter reduces a caller ID's quota within the transaction
pool to some fraction of the overall pool size. We've found it useful for
tightening the blast radius when a single client starts to have long
transactions for whatever reason.

We noticed in a development enviornment that the tx limiter was preventing
any new transactions for a particular caller ID even though no transactions
for that user were already in progress.

I'm not too familiar with this code, although it seems a bit brittle. In this PR
I've tried to make sure that if there's an error grabbing a connection from the
connection pool or if a connection is killed by the transaction killer, tx limiter
quota is restored.

I'd love for anyone who is to take a close look at this.

It seems possible that killed transactions were also previously not logged, and
this might fix that too?...

See https://github.com/vitessio/vitess/issues/6727

Signed-off-by: David Weitzman <dweitzman@pinterest.com>