Fix error around breaking of multistatements
When a multistatement query is received, any errors should
abort the execution of the remaining queries.

The `execQuery` and `handleNextCommand` were returning an error,
but not actually using the error value - just checking if it was nil or not.

We need to be able to know on the outside of `execQuery` if an error occured and if it was an error we need to close the connection for or if it was a simple execution error.

Signed-off-by: Andres Taylor <andres@planetscale.com>