Improvements to queryz
* Adds supports for multiple sorters: counts per query, mysql time per query and
rows per query.
* Also exposes a way to reset the cache, so the counters can have a fresh start.
This could be useful to debug issues with unexpected changes in query patterns.
For tablets that have been running for a long time, the history outweights
potentially new queries that are being introduced. Ideally we would like some
sort of graph that shows this is in a more obvious way, but the reset is an easy
change that gives significant benefits as well.

Signed-off-by: Rafael Chacon <rafael@slack-corp.com>