tablet streaming health fix: never silently skip health state changes
Today, if vtgate is slow to receive health updates from a tablet the tablet may
simply throw away the most recent status update instead of buffering it.

Since there's no periodic forced state push and discarded state is the most
recent one, this creates a possibility that a gate can permenently and incorrectly
believe that a tablet is not serving. You could have a healthy master tablet and a
gate that doesn't believe it's healthy and the only fix today would be to do an
operation and restart the gate.

The fix in this PR is to cancel the streaming health RPC if the buffer overflows rather
than discarding the most recent state update. Whenever the gate or its network connection
gets unwedged it'll realize that the streaming health RPC is canceled and open up a new
streaming health RPC where it will get a recent, accurate state.

Fixes https://github.com/vitessio/vitess/issues/5445

Signed-off-by: David Weitzman <dweitzman@pinterest.com>