vttablet: more robust query kill
The previous query kill waited for mysql to return an error after
initiating the kill. Sometimes, mysql can take a long time to kill
queries. In this new code, we proactively close the client side
connection. This will cause the execute to return immediately.

We also return a custom non-connection error that will prevent
the Exec function from unnecessarily retrying the query.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>