Add client command line implementation
We bring in cobra to manage the CLI for us; it's the most structured one
Vitess has had yet. This also forces us to do _something_ about the
various project-global flags, but it's a compromise - essentially we
just blindly apply all of them to the root of our command tree.

Signed-off-by: Andrew Mason <amason@slack-corp.com>