Do not swap out the already-populated flagset for the parse hack
It turns out that some vitess packages make calls to `flag.Lookup` to
find flag values, and if we swap out the original flagset, these all
fail, and in some cases, panic. I encountered this while attempting to
move vttablet to cobra in a separate branch, and the panic came from
`logutil.PurgeLogs()` needing a `log_dir` flag ([ref][logutil_ref]).

[logutil_ref]: https://github.com/vitessio/vitess/blob/20a545533bc94d6f85e1682a6539e93c32a82486/go/vt/logutil/purge.go#L106-L109

While this appears to actually not matter for the vtctldclient, since
the code with the potential bug is able to run without issue, I don't
want to commit a potentially-risky flag shim into the codebase that may
be copied by others.

Signed-off-by: Andrew Mason <amason@slack-corp.com>