Migrate InitShardMaster=>InitShardPrimary to VtctldServer, update examples
There are no functional changes in this commit, just a pure port with
the most minor changes to ensure identical behavior (like the callback
logger).

I also updated our `proto` target to assume a standard protoc
installation (which installs a stdlib of well-known types into
/usr/local/include). We can decide if this is okay to make a
requirement, or if we should try to make this more portable in PR
review.

I also updated the local example to use the vtctldclient port, and went
through the full example with no issues.

Signed-off-by: Andrew Mason <amason@slack-corp.com>