keep-heartbeat-reader-running-on-master
From Slack discussion: https://vitess.slack.com/archives/C0PQY0PTK/p1606781256023700

"...the issue is, when you stop the ticker under that condition, the lag
metric gets frozen at a value, but continues to be exported.  This leads
to problematic metric values wherein e.g. you have a replica constantly
reporting 10 seconds of lag, and there’s no simple way in prometheus to
determine that this is the lag of the replica that is the master at that
point in time
"

Signed-off-by: Andrei Sura <andrei.sura@sharpspring.com>