Bypass MySQL Cert issue, and update ETCD api
This sets the flag to ignore 509 validation errors caused by certs not
setting a SAN and only using a CN. The default MySQL certs don't set
either to a valid hostname.

Newer versions of ETCD don't respond on the v2 endpoint so this replaces
that check with the /version endpoint which will always respond.

Signed-off-by: D.K <dan.kozlowski@gmail.com>