executor: do not lookup in cache twice
There is no need to lookup ta given plan in the cache twice (in its
normalized and non-normalized representation for its key): if the plan
is normalized, it'll be stored into the cache in its normalized form. If
it's not normalized, it'll be stored in its original form. Either way,
the initial lookup with its non-normalized form is redundant.

The raw `sql` content of a query only changes if the query has been
normalized. In cases where it hasn't, there is no need to lookup the
same key twice on cache

Signed-off-by: Vicent Marti <vmg@strn.cat>