mariadb multiple version secure user support
MariaDB-10.4 uses mysql.user as a VIEW so the old mechanism of securing
won't work.

To replace it a combination DROP USER for the default created users
is used.

Using the MariaDB-10.2, the current minimum upstream supported version,
DROP USER ''@ @@hostname and DROP USER root@ @@hostname are used.

10.3+ uses a more thorough cursor on the mysql.user table to remove users.
A more elaborate use of CURSOR would be able to use the same on 10.2
as this supports EXECUTE IMMEDIATE.

The previous behavior of deleting to mysql.user is limited to MySQL-5.7+
(and not MariaDB). MySQL-5.6 is EOL In February 2021 by all
accounts (https://lefred.be/content/mysql-5-6-eol-is-february-2021/)
so this is a few days pre-emptive.

Signed-off-by: Daniel Black <daniel@mariadb.org>