Extracting some reparent code from the wrangler struct to pure functions, with tests
This is for later code sharing between the legacy and new vtctld APIs.

This commit also contains a bugfix in `FindValidEmergencyReparentCandidates` / `(*Wrangler).findValidReparentCandidates`
that causes a non-deterministic panic, depending on what order the
`replicationStatusMap` is first iterated over in. The panic happens as
follows:

Assume you have 2 tablets in the status map. One tablet has a GTID-based
relay log position, and the other has a non GTID-based relay log position.

If the GTID tablet comes first, then `isGtidBased` gets set to a pointer
to true, and when we come to the second, non-GTID tablet, we fall through
to the 3rd case in the switch.

If the non-GTID tablet's position is zero, then we return an error, and
everything's fine. _However_, if it's non-zero _and_ non-GTID, then we
successfully make it through the first loop, and later on, we try to
`FindErrantGTIDs`, because `*isGtidBased` is true. This is where the panic
happens, because we give it a mix of `Mysql56GTIDSet` and
non-`Mysql56GTIDSet` positions.

<hr>

The fix here is to iterate through the full status map, and not break
out early. We track whether we've seen _any_ GTID-based or non-GTID
based positions, and fail early accordingly.

Signed-off-by: Andrew Mason <amason@slack-corp.com>