[ERS] confirm at least one replica succeeded to `SetMaster`, or fail
This introduces two background goroutines and contexts derived from
`context.Background()` to poll and signal for (1) at least one replica
finished successfully and (2) all replicas finished, regardless of
status, respectively.

The overall `promotePrimary` function returns error conditions based on
which of those contexts gets signaled first. We still fail if
`PromoteReplica` fails, which gets checked first, but this covers the
case where we're not running with semi-sync, and none of the replicas
are able to reparent even if the new primary is able to populate its
reparent journal (in the semi-sync case, if 0 replicas succeed to
`SetMaster`, than the primary will fail to `PromoteReplica`, so this was
already covered there).

Fixes #7480

Signed-off-by: Andrew Mason <amason@slack-corp.com>