mysql/server: add PreHandleFunc hook for connections
This PR adds a new `PreHandleFunc` field to the Listener struct. The
goal of this hook is to inspect the connection or add a custom
TLS termination. Because  MySQL does TLS termination differently, it'll
allow us to terminate first and then let the handler continue processing
the connection according to the MySQL protocol.

I iniatally wanted it to make a part of the `Handler` interface, but
that would be breaking change. If we're ok breaking it, I can make it
part of the handler.

Signed-off-by: Fatih Arslan <ftharsln@gmail.com>