vstreamer: do not reload twice
There is a bad merge in this code that is causing `ReloadAt` to be
called two times in a row. Previously, the schema was _always_ being
reloaded when parsing a DDL statement. @sougou changed this so the
schema only reloads when the statement must be sent upstream, but then
Rohit brought back the functionality of _always_ reloading the DDL
without removing the case where we were already reloading for sent
statements.

Signed-off-by: Vicent Marti <vmg@strn.cat>